Refactor login form
Added login form validation with Formil and Yup
Added free font-awesome icons
Some style fixes and layout improvements
