--- src/components/LoginView.js
@@ -1,30 +1,0 @@
-import React from 'react'
-import { connect } from 'react-redux'
-import LoginForm from './LoginForm'
-import { Container, Row, Col, Image } from 'react-bootstrap'
-
-const LoginView = () => {
-
-	return (
-		<Container className="py-4">
-			<Row className="d-flex justify-content-center align-items-center">
-				<Col xs={12} sm={8} md={6} lg={4} className="py-4 text-center">
-					<Image src="img/schoolLogo-transparent.png" className="responsive-image-half"/>
-				</Col>
-				<Col xs={12} sm={8} md={6} lg={4}>
-					<LoginForm />
-				</Col>
-			</Row>
-		</Container>
-	)
-}
-
-const mapStateToProps = (state) => {
-	return {
-		user: state.user
-	}
-}
-
-export default connect(
-	mapStateToProps
-)(LoginView)
