+++ src/css/style.css
@@ -200,9 +200,15 @@ html {
   width: 100%;
 }
 
-.responsive-image-half {
+.responsive-image-fraction {
   height: auto;
-  width: 50%;
+  width: 75%;
+}
+
+.image-align-helper {
+  display: inline-block;
+  height: 100%;
+  vertical-align: middle;
 }
 
 .primary-color-shadow {
@@ -242,6 +248,10 @@ html {
   height: 100%;
   width: 100%;
   position: absolute;
+}
+
+.elevated-z-index {
+  z-index: 99;
 }
 
 /*# sourceMappingURL=style.css.map */
