+++ src/sass/style.sass
@@ -25,9 +25,14 @@ html
 	height: auto
 	width: 100%
 
-.responsive-image-half
+.responsive-image-fraction
 	height: auto
-	width: 50%
+	width: 75%
+
+.image-align-helper
+	display: inline-block
+	height: 100%
+	vertical-align: middle
 
 .primary-color-shadow
 	border-radius: .3rem
@@ -59,3 +64,6 @@ html
 		height: 100%
 		width: 100%
 		position: absolute
+
+.elevated-z-index
+	z-index: 99
